git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/sidelnik@19864 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/src/chplsys.c
@@ -60,7 +60,7 @@ uint64_t chpl_bytesPerLocale(void) {
 }
 
 
-int32_t chpl_numCoresOnThisLocale(void) {
+int64_t chpl_numCoresOnThisLocale(void) {
 #ifdef NO_CORES_PER_LOCALE
   return 1;
 #elif defined __APPLE__
