+++ test/extern/sungeun/classes/extern_defs.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 
 typedef struct C_s {
-  int x;
+  int64_t x;
 } _C, *C;
 
-void foo(C, int);
+void foo(C, int64_t);
